requirements.txt
streamlit
spacy
nltk
python-docx
PyPDF2
pandas
numpy
transformers
torch
scikit-learn
